django-practice/pyproject.toml
target-version = "py313"

[project]
name = "django-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-environ>=0.11.2",
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "django-extensions>=3.2.3",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=4.4.6",
    "pre-commit>=4.0.1",
    "ruff>=0.8.2",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
ignore = ["E501", "E741"]
